mi-mctp: Fix compile bug in test/mi-mctp.c when !HAVE_LINUX_MCTP_H
The test/mi-mctp.c fails to compile in certain cases where
HAVE_LINUX_MCTP_H is false.  There's a workaround for this condition
in src/mi-mctp.c but not here.  This change makes a common header and
includes it in both cases.
